렌트 머지

+++ KingGgang_project/src/main/java/com/teamb/controller/MyController.java
@@ -33,10 +33,11 @@ import com.teamb.model.MemberDTO;
 import com.teamb.service.MemberMapper;
 import com.teamb.service.MyMapper;
 
-
+  
 
 /*
 이	   름 : MyController.java
+
 
 개  발   자 : 원세호
 설	   명 : 마이페이지 컨트롤러
@@ -82,29 +83,36 @@ public class MyController {
 		//String password = String.valueOf(session.getAttribute("password"));
 		String id =  req.getParameter("id");
 		String password = req.getParameter("password");
+		System.out.println(id);
+		MemberDTO dto = myMapper.getMemberNo(id);
 		
-		MemberDTO dto = myMapper.getMember(id);
-		String filename = dto.getProfile_name(); 
-		File file = new File(upLoadPath,filename);
-		
-		int res = myMapper.deleteMember(id,password);
+		int memberNum = dto.getMemberNum();
+		System.out.println(memberNum);
+		int res = myMapper.deleteMember(id,password,memberNum);
 		String msg = null, url=null;
-		if (res>0) {
-			if(file.delete()) {
-			url = "home.my";
-			msg = "delete successed";
-			}else {
-				url = "home.my";
-				msg = "delete successed but image is remained";
-			}
-		}else {
-			url = "home.my";
-			msg = "delete failed";
-		}
+		if (res > 0) {
+	         if (dto.getProfile_name() == null) {
+	            url = "memberList.mem";
+	            msg = "delete successed";
+	         } else {
+	            String filename = dto.getProfile_name();
+	            File file = new File(upLoadPath, filename);
+	            if (file.delete()) {
+	               url = "memberList.mem";
+	               msg = "delete successed";
+	            } else {
+	               url = "memberList.mem";
+	               msg = "delete successed but image is remained";
+	            }
+	         }
+	      } else {
+	         url = "memberList.mem";
+	         msg = "delete failed";
+	      }
 		
 		req.setAttribute("url", url);
 		req.setAttribute("msg", msg);
-		return "my/alert";
+		return "message";
 	}
 	
 	@RequestMapping("/update.my")
@@ -115,7 +123,7 @@ public class MyController {
 	}
 	@RequestMapping(value = "/updateForm.my")
 	public ModelAndView updateMem(@RequestParam String id) {
-		MemberDTO dto =  myMapper.getMember(id);
+		MemberDTO dto =  myMapper.getMemberNo(id);
 		
 		ModelAndView mav = new ModelAndView("my/updateForm");
 		mav.addObject("getMember", dto);
@@ -168,7 +176,7 @@ public class MyController {
 	}
 	req.setAttribute("url", url);
 	req.setAttribute("msg", msg);
-	return "my/alert";
+	return "message";
 	}
 	
 	@RequestMapping("/home.my")
